Refine sidebar CSS
The old setup required a hacky height for the `sticky` sidebar which
made the sidebar looked buggy. It also forced an awkward scrollbar
position for the main content which felt confusing.

This new approach is slightly better, but the one downside is that it
requires a fixed navbar. The consolation is that we can un-fix this
when the sidebar is hidden.