Add TextBlock table
We need some version of this for cross-referencing, e.g. when linking
parse data to a specific verse. While we can implement some version of
this by using TextSection.xml alone, there are there are fewer integrity
guarantees.

In the future, we might use some kind of XML database. But for now, this
should work just fine.