Remove unused PyMuPDF from requirements.txt
It was hard to set up in my local dev environment (first complained
about swig not being installed, then fitz.h header): turns out it's
not even used, so simpler to just remove it.

Also updated "python" to "python3" in the installation instructions,
as it doesn't exist on my system (which is not unusual, see PEP 394):
he other invocations are ok as they happen inside the venv where it
("python") is available.