Add admin view
This view lets us delete interesting objects on Ambuda. The admin setup
is quite clunky, but we can add more power to it soon.

NOTE: for now, the admin view expects an admin user named `akprasad`.
In the future, any user with an admin role will able to use the admin
panel.