Introduce basic CI for Python tests
Address #16.

Runs on pull requests, pushes to main, and manual invocations.

The badge will appear as a broken link until this is run on main.

For now its just the python unit tests, but can serve as a model for
addition additional lints, code coverage reports, etc.

View the "checks" tab under this test PR to see the results:

https://github.com/theFestest/ambuda/pull/1