Add lint to CI
Rationale for two separate jobs: two jobs is faster (parallelize), and it also helps
you quickly iterate -- i.e., if you are trying to add some commits to fix a test, you
don't have to make sure the lint passes for each PR, you can just fix the lint at the
very end once the PR is ready. It can be frustrating sometimes to make changes to
run tests on CI, but then have to re-run the whole pipeline just because the lint failed.