Add cleaner config setup and comments
All config values are defined and documented on a `BaseConfig` from
which all config schemas inherit. Further, I added extra validation
checks to ensure that our configs are well-formed.

The one thing I dislike here is that `_env` will happily return None if
an environment variable can't be found. It has this behavior (as opposed
to just failing) because otherwise, class initialization will fail if a
value is defined and we won't be able to import the module.

The workarounds to this that I can think of are clumsy, but we can
revisit this code later.