Add explicit CC0 statement to page edit form
This makes an implicit point explicit: our contributions, to the extent
that they have copyright in different jurisdictions, should be dedicated
to the public domain. My understanding is that the digital text that we
produce on Ambuda will thus have the same access permissions as the
original book on which it is based.