Add project creator_id, created_at, and updated_at
`creator_id` tracks the project creator for better visibility into the
overall effort. This is nullable for now because we need to manually
migrate the production setup first.

`created_at` is likewise for better visibility, and it can show proofers
which projects have been in progress the longest.

`updated_at` is not useful now, but in the future, it can show proofers
which projects have been changed most recently.

Test plan: unit tests, ran `upgrade` and `downgrade` and `upgrade`
again.