Use Celery project uploader in cli.py
Now users can run `./cli.py create-project` to create a new proofing
project from a PDF. This command uses the same underlying logic that we
use on Celery, but it doesn't require any Celery or Redis setup.

Test plan: Unit tests and ran `/cli.py create-project` on a local PDF. I
also uploaded a PDF through the uploader on the devserver to confirm
that the Celery code still works as planned.