Add page to see all proofing threads at once
Our discussion threads don't have a notifications mechanism, which
greatly reduces their usefulness. As a quick workaround, create a single
link that shows all discussion across all projects.

The implementation is a litle hacky, but we can refine if it this
becomes a popular feature.

Test plan: unit tests and opened the new page on the dev server.