Optimize proofing.index query
The page is loading slowly for me in prod. Upon debugging, I found that
our current prod setup runs N+2 queries, where N is the number of
projects:

- 1 to fetch all projects
- N to fetch pages for each project
- 1 to fetch page statuses.

Since we use sqlite, this isn't as bad as it seems. But even so, I
thought there was room to improve this query by fetching our data in one
trip and dropping extraneous fields.

Test plan: unit tests and ran the query on dev.