Bugfix: Make it possible to save document started with an empty textarea.
The problem with setting `required` is that if we start with an empty textarea,
create an editor state from it, add text to it, etc., then (because the textarea
is still empty at the time of the submit event) the submit event (which would
populate the textarea) doesn't even fire (because the textarea is empty), which
prevents anything from being submitted.

An alternative fix would be to populate the textarea more frequently, but this
seems wasteful (and we'll probably move away from the textarea soon anyway).