Add basic Sanskrit interface
This commit adds basic support for a Sanskrit interface, which the user
can select by clicking a link in the page footer. The existing
translation is mainly a proof-of-concept and translates only a small
portion of the site.

We implement this logic by storing the user's locale preference as part
of the session.

Test plan: tried it on the devserver.