Fix bug: single-section texts show "all" in nav
A *single-section* text is a text with no major divisions -- no kaandas,
no sargas, etc. Internally, we represent such texts as texts with a
single section with the slug `"all"`.

Our current reader UI displays this `"all"` slug as if it's a valid
identifier, which is unsightly. So as a fix, show the section slug only
if the text has multiple sections.

Test plan: clicked around on dev