[proofing] Add filter field for project list
Our proofers have requested more tools to search and manage the list of
available projects. This commit adds a simple text form that performs
case-insensitive substring matching against the list of projects.

The implementation itself is quite clumsy since it manages state in both
JS data and in the order and attributes of HTML elements. I think this
messy approach is acceptable given that this is a new feature and the
actual logic is quite short. If our proofers ask for other filtering
options in the future, we can clean this up.

Test plan: tried it on dev.