[admin] Add new templates
Add new templates to the admin interface to improve the look and feel.

As Ambuda grows, it will be useful to grant certain users access to CRUD
functionality for certain database objects. One example is our planned
list of projects available for sponsorship.

Rather than roll custom interfaces for each of these objects, it's
simpler to reuse Flask-Admin.