[proofing] Hide OCR button for pages with edits
We want users to use OCR if they need to, but we don't want them to
re-run OCR on an edited page. We also have less need for the prominent
"Run OCR" button now that we have batch OCR.

New UX: If a page has no edits, show the user a help box prompting them
to run OCR. If a page has edits already, hide the help box. OCR is
always available under the `Tools` menu if it's really needed.

Test plan: unit tests and tried it on dev.